Added missing tests for users variables
