+++ test/util/test_expression.py
@@ -9,8 +9,12 @@ import numpy as np
 import numpy.testing as npt
 import xarray as xr
 
+from xcube.core.maskset import MaskSet
+from xcube.core.new import new_cube
 from xcube.util.expression import compute_array_expr
 from xcube.util.expression import compute_expr
+from xcube.util.expression import new_dataset_namespace
+from xcube.util.expression import split_var_assignment
 from xcube.util.expression import transpile_expr
 
 
@@ -334,3 +338,37 @@ class TranspileExprTest(unittest.TestCase):
             "np.logical_or(floating_cyanobacteria == 1, chl_pitarch > 500)))",
             transpile_expr(expr),
         )
+
+
+class HelpersTest(unittest.TestCase):
+
+    def test_new_dataset_namespace(self):
+
+        ds = xr.Dataset(
+            dict(
+                A=xr.DataArray([[0, 1], [2, 3]], dims=("y", "x")),
+                B=xr.DataArray([[1, 2], [3, 4]], dims=("y", "x")),
+                C=xr.DataArray(
+                    [[1, 0], [0, 1]],
+                    dims=("y", "x"),
+                    attrs={"flag_meanings": "on off", "flag_values": "0, 1"},
+                ),
+            )
+        )
+
+        ns = new_dataset_namespace(ds)
+        self.assertEqual({"A", "B", "C", "PI", "NaN", "np", "xr"}, set(ns.keys()))
+        self.assertIsInstance(ns["A"], xr.DataArray)
+        self.assertIsInstance(ns["B"], xr.DataArray)
+        self.assertIsInstance(ns["C"], xr.DataArray)
+
+        ns = new_dataset_namespace(ds, use_mask_sets=True)
+        self.assertEqual({"A", "B", "C", "PI", "NaN", "np", "xr"}, set(ns.keys()))
+        self.assertIsInstance(ns["A"], xr.DataArray)
+        self.assertIsInstance(ns["B"], xr.DataArray)
+        self.assertIsInstance(ns["C"], MaskSet)
+
+    def test_split_var_assignment(self):
+        self.assertEqual(("A", None), split_var_assignment("A"))
+        self.assertEqual(("A", "B"), split_var_assignment("A=B"))
+        self.assertEqual(("A", "B + C"), split_var_assignment(" A = B + C "))
